+++ spec/integration/outputs/no_es_on_startup_spec.rb
@@ -12,13 +12,13 @@ describe "elasticsearch is down on startup", :integration => true do
                                            "template_overwrite" => true,
                                            "hosts" => get_host_port(),
                                            "retry_max_interval" => 64,
-                                           "retry_initial_interval" => 2
+                                           "retry_initial_interval" => 2,
+                                           'ecs_compatibility' => 'disabled'
                                        })
   }
 
   before :each do
     # Delete all templates first.
-    require "elasticsearch"
     allow(Stud).to receive(:stoppable_sleep)
 
     # Clean ES of data before we start.
@@ -33,7 +33,9 @@ describe "elasticsearch is down on startup", :integration => true do
   end
 
   it 'should ingest events when Elasticsearch recovers before documents are sent' do
-    allow_any_instance_of(LogStash::Outputs::ElasticSearch::HttpClient::Pool).to receive(:get_es_version).and_raise(::LogStash::Outputs::ElasticSearch::HttpClient::Pool::HostUnreachableError.new(StandardError.new, "big fail"))
+    allow_any_instance_of(LogStash::Outputs::ElasticSearch::HttpClient::Pool).to receive(:get_es_version).and_raise(
+        ::LogStash::Outputs::ElasticSearch::HttpClient::Pool::HostUnreachableError.new StandardError.new("TEST: before docs are sent"), 'http://test.es/'
+    )
     subject.register
     allow_any_instance_of(LogStash::Outputs::ElasticSearch::HttpClient::Pool).to receive(:get_es_version).and_return(ESHelper.es_version)
     subject.multi_receive([event1, event2])
@@ -43,7 +45,9 @@ describe "elasticsearch is down on startup", :integration => true do
   end
 
   it 'should ingest events when Elasticsearch recovers after documents are sent' do
-    allow_any_instance_of(LogStash::Outputs::ElasticSearch::HttpClient::Pool).to receive(:get_es_version).and_raise(::LogStash::Outputs::ElasticSearch::HttpClient::Pool::HostUnreachableError.new(StandardError.new, "big fail"))
+    allow_any_instance_of(LogStash::Outputs::ElasticSearch::HttpClient::Pool).to receive(:get_es_version).and_raise(
+        ::LogStash::Outputs::ElasticSearch::HttpClient::Pool::HostUnreachableError.new StandardError.new("TEST: after docs are sent"), 'http://test.es/'
+    )
     subject.register
     Thread.new do
       sleep 4
@@ -55,4 +59,20 @@ describe "elasticsearch is down on startup", :integration => true do
     expect(r).to have_hits(2)
   end
 
+  it 'should get cluster_uuid when Elasticsearch recovers from license check failure' do
+    allow_any_instance_of(LogStash::Outputs::ElasticSearch::HttpClient::Pool).to receive(:get_license).and_raise(
+        ::LogStash::Outputs::ElasticSearch::HttpClient::Pool::HostUnreachableError.new StandardError.new("TEST: docs are sent"), 'http://test.es/_license'
+    )
+    subject.register
+    Thread.new do
+      sleep 4
+      allow_any_instance_of(LogStash::Outputs::ElasticSearch::HttpClient::Pool).to receive(:get_license).and_call_original
+    end
+    subject.multi_receive([event1, event2])
+    @es.indices.refresh
+    r = @es.search(index: 'logstash-*')
+    expect(r).to have_hits(2)
+    expect(subject.plugin_metadata.get(:cluster_uuid)).not_to be_empty
+    expect(subject.plugin_metadata.get(:cluster_uuid)).not_to eq("_na_")
+  end if ESHelper.es_version_satisfies?(">=7")
 end